feature: Email validation on string or array of string written in TS
Move the array to its own file.

Import EmailValidation in the input and create a basic button for validating. 

The email input is now of type email, I've only checked Chrome but it makes a label on hover that declares an email to be invalid even when the custom regexp returns true. It does not affect the validation in any way.