chore: Updare tsconfig.json and remove redundant properties
(styleguide: https://udacity.github.io/git-styleguide/ )
body(optional):
Explain the problem that this commit is solving. Focus on why you
are making this change as opposed to how (the code explains that).
Are there side effects or other unintuitive consequenses of this
change? Here's the place to explain them.

footer(optional):
Resolves: #123
See also: #456, #789